🌡️🥋️ ↝ [SSG-90 SSG-91]: Confirming that post cards are less redundant now (also discovered a new bug, yay)
