+++ app/tests/page.tsx
@@ -2,10 +2,6 @@
 
 import React, { useState } from "react";
 import StarnetLayout from "@/components/Layout/Starnet";
-import { MiningComponentComponent } from "@/components/(scenes)/mining/mining-component";
-import Greenhouse from "@/page.test";
-import BigMap from "@/components/(scenes)/planetScene/bigMap";
-import DiscoveriesPage from "@/content/Classifications/minimalDiscoveries";
 import DailyMinorPlanetMissions from "@/components/Structures/Missions/Astronomers/DailyMinorPlanet/DailyMinorPlanet";
 // import { TopographicMap } from "@/components/topographic-map";
 
@@ -14,10 +10,7 @@ export default function TestPage() {
         // <StarnetLayout>
           <>
           <DailyMinorPlanetMissions />
-          <DiscoveriesPage />
               {/* <Greenhouse /> */}
-              {/* <BigMap /> */}
-              {/* <DiscoveriesPage /> */}
               {/* <MiningComponent /> */}
           </>
         // {/* </StarnetLayout> */}
