+++ components/Structures/Missions/Astronomers/PlanetHunters/VoteForm.tsx
@@ -124,5 +124,3 @@ const MissionFour: React.FC = () => {
     </div>
   );
 };
-
-export default MissionFour;
